fix(recents): remove known addresses from recipients
